+++ CHANGES.rst
@@ -17,6 +17,15 @@ Changelog
 - Fix upgrade step error #139 
   [mamico]
 
+- Add flag `auto_confirm_manager` that allows to automatically confirm bookings created by Booking Managers.
+  [cekk]
+
+- Booking Managers bypass futureDays checks when create new bookings.
+  [cekk]
+
+- Remove duplicated method in booker: create. Now there is only `book`.
+  [cekk]
+
 
 2.2.5 (2023-12-04)
 ------------------